// types.ts
export interface ZipFileEntry {
  name: string;
  path: string;
  content: string;
  isFolder: boolean;
  size: number;
}